+++ CHANGELOG.md
@@ -6,9 +6,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [0.0.1] - 2019-11-03
 ### Added
 - `Priconneyomi` phrases are now gray-scaled if it is Kaya's turn
+- Latest selection will now be displayed at the top
 ### Changed
 - Word list is now in an actual list rather than row-per-category
 - Adjusted Open Graph icon size
+- `Priconneyomi` phrases will no longer appear at the start
 ### Fixed
 - Fixed a lot of issues regarding characters, i.e. `ゅ` was used instead of `ゆ`